requirements.txt
amqp==5.0.6
APScheduler==3.6.3
asgiref==3.4.1
billiard==3.6.4.0
cachetools==4.2.2
celery==5.1.2
certifi==2021.5.30
charset-normalizer==2.0.6
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.6
django-admin-sortable2==1.0.2
django-author==1.0.2
django-import-export==2.6.1
django-import-export-celery==1.1.4
django-nested-inline==0.4.4
et-xmlfile==1.1.0
idna==3.2
kombu==5.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
Pillow==8.3.2
prompt-toolkit==3.0.20
psycopg2==2.9.1
python-dateutil==2.8.2
python-decouple==3.4
python-dotenv==0.19.0
python-telegram-bot==13.7
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.26.0
setuptools-git==1.2
six==1.16.0
sqlparse==0.4.1
tablib==3.0.0
tornado==6.1
tzlocal==3.0
urllib3==1.26.7
uWSGI==2.0.19.1
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.3.0
xlrd==2.0.1
xlwt==1.3.0
